requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.40
psycopg2==2.9.10
psycopg[binary]
python-dotenv==1.1.0
pytest==8.3.1

# para crear los certificados de los alumnos
weasyprint==65.1